+++ datasources/src/main/java/com/hopovo/mobile/android/prepexam/datasources/exercise/remote/model/ExerciseResponse.kt
@@ -3,5 +3,5 @@ package com.hopovo.mobile.android.prepexam.datasources.exercise.remote.model
 /**
  * Representation for a [ExerciseResponse] fetched from the API
  */
-data class ExerciseResponse(val id: Long, val description: String, val title: String,
+data class ExerciseResponse(val id: Int, val description: String, val title: String,
                             val image: String, val level: String, val time: String)
